Address ctype data type warnings in Cygwin
Ctype macros such as isprint() and isdigit() want int parameters, passing
char issues a warning about array subscripts being char.

+++ src/loaders/common.c
@@ -27,7 +27,7 @@ char *copy_adjust(char *s, uint8 *r, int n)
 	strncpy(s, (char *)r, n);
 
 	for (i = 0; s[i] && i < n; i++) {
-		if (!isprint(s[i]) || ((uint8)s[i] > 127))
+		if (!isprint((int)s[i]) || ((uint8)s[i] > 127))
 			s[i] = '.';
 	}
